refactor: api 이름 변경
