+++ src/api/your-fights.tsx
@@ -1,13 +1,29 @@
 import { app } from "../shared/FireBase";
 import {
+  addDoc,
   collection,
   getFirestore,
   getDocs,
-  doc,
   query,
   orderBy,
 } from "firebase/firestore";
-import { useState, useEffect } from "react";
+import { getAuth } from "firebase/auth";
+import getRandomNickname from "./get-random-nickname";
+
+const db = getFirestore(app);
+const auth = getAuth();
+const uid = auth.currentUser?.uid ?? "";
+
+interface Props {
+  content: string;
+  optionList: option[];
+  likes: LikesType;
+}
+
+interface option {
+  optionValue: string;
+  votes: number;
+}
 
 export type LikesType = {
   like: number;
@@ -39,7 +55,30 @@ interface DocType {
 export interface IOptionListType extends Array<OptionListType> {}
 export interface IDocType extends Array<DocType> {}
 
-const db = getFirestore(app);
+export const postYourFightsForm = async ({
+  content,
+  optionList,
+  likes,
+}: Props) => {
+  try {
+    const nickname = await getRandomNickname();
+
+    await addDoc(collection(db, "yourFights"), {
+      createdAt: Date.now(),
+      user: {
+        uid,
+        nickname,
+      },
+      data: {
+        content,
+        optionList,
+        likes,
+      },
+    });
+  } catch (e) {
+    console.error("Error adding document: ", e);
+  }
+};
 
 export const getYourFightsOrderByDate = async () => {
   try {
