+++ src/pages/your-fights/components/OrderBy.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { useTheme } from "@mui/system";
 import CheckIcon from "@mui/icons-material/Check";
 import styled from "@emotion/styled";
